+++ packages/Angular/Generic/forms/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @memberjunction/ng-forms
 
+## 6.1.0-edge.4
+
+### Patch Changes
+
+- @memberjunction/ai-core-plus@6.1.0-edge.4
+- @memberjunction/ng-markdown@6.1.0-edge.4
+- @memberjunction/ng-ui-components@6.1.0-edge.4
+
 ## 6.1.0-edge.3
 
 ### Patch Changes