+++ packages/CLICore/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @memberjunction/cli-core
 
+## 6.1.0-edge.4
+
+### Patch Changes
+
+- Updated dependencies [4586215]
+- Updated dependencies [a5f92d2]
+  - @memberjunction/global@6.1.0-edge.4
+
 ## 6.1.0-edge.3
 
 ### Patch Changes